[CI skip] Fix vouch generator docs scripts Also: - Move doc generator logic to a separate file - Reimplemented sidebar generation function
